chore: upgrade gradle; reorganize receiver package
