+++ dependencies.gradle
@@ -13,7 +13,7 @@ def deps = [:]
 
 def versions = [:]
 versions.plugin_android_maven = "2.1"
-versions.plugin_build_tools = '4.1.0-alpha09'
+versions.plugin_build_tools = '4.2.0-alpha02'
 versions.plugin_google_services = "4.3.0"
 versions.plugin_fabric = '1.+'
 versions.plugin_firebase = '1.2.0'
