+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,6 @@
-#Sat May 16 13:01:08 EDT 2020
+#Mon Jul 06 22:57:50 EDT 2020
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.4-rc-2-all.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.5-rc-1-bin.zip
