+++ library/src/main/java/dev/bmcreations/scrcast/recorder/receiver/RecordingNotificationReceiver.kt
@@ -1,9 +1,10 @@
-package dev.bmcreations.scrcast.recorder
+package dev.bmcreations.scrcast.recorder.receiver
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import dev.bmcreations.scrcast.recorder.Action
 
 class RecordingNotificationReceiver : BroadcastReceiver() {
